Change Appearance of Main Activity
1) Completely switched up the colour scheme of the main activity including the overall background, the grid background, and the colours associated with P1 and P2 respectively to fit in clearly.

2) Took care of the few edge cases that needed to be handled from the initial commit.

At this point, the pass-and-play Tic Tac Toe game is working completely as expected, and I am satisfied with the current default colour scheme. Now, I am looking to add additional activities to the app, namely a landing activity and a settings activity. I expect the landing activity to not have much to it, apart from two buttons for either playing (continuing to the main activity) or exiting the app, and a button in the top-right corner leading to the settings activity. I intend on providing the user with the ability to customize the look and appearance of the app through the settings activity. More specifically, I will likely create a couple different colour themes for the app, which the user will be able to choose from in the settings activity; by doing this, I will give users some freedom in terms of layout customization but at the same time ensure that each of the layout options works out nicely (in terms of the colour combinations). Although I haven't added any sounds to the button clicks or anything for that matter, I do intend on adding this at some point, and consequently also an option to toggle sound in the settings activity.